LayoutTests/imported/w3c/web-platform-tests/workers/semantics/multiple-workers/008-expected.txt

CONSOLE MESSAGE: TypeError: null is not an object (evaluating 'w1.port.onmessage = function(e) {
  e.ports[0].postMessage(2);
}')

FAIL messagechannel in shared worker null is not an object (evaluating 'w1.port')
